# Vectorize string parsing