I've implemented [your suggestion here](https://github.com/serde-rs/json/pull/1160#pullrequestreview-2203573484), and I need a bit of your input regarding my approach.

What we *really* want is to find the first character s.t. `ch == b'"' || ch == b'\\' || ch < 0x20`. memchr3 supports searches of kind `ch == a || ch == b || ch == c`, which is almost what we need. SIMD comparisons for equality and non-equality are not significantly different performance-wise or implementation-wise, so memchr could theoretically help us out. I've filed https://github.com/BurntSushi/memchr/issues/157, but chances are this isn't going to be implemented anytime soon.

The second best thing -- and I want to articulate that this is *not* the more performant solution -- is to first use SIMD to scan for `"` or `\` and then find escapes in an ad-hoc way. However, it turns out that `ch < 0x20` in a hot loop is not any faster than `ESCAPES[ch]`, on benchmarks at least. The reason is likely that the table is brought into L1 cache and the string isn't, so the speed of string iteration shadows the concerns about table access.

This forces us to use SIMD in serde-json *directly*. Luckily, there's a way to implement what we need in an architecture-agnostic way, *and* it's faster than using intrinsics on my benchmarks. This unfortunately increases code complexity a bit and requires an `#[inline(always)]` annotation, but this is what we get in return:

```
                    old (MB/s)  new (MB/s)  perf increase
canada dom          289         289         0%
canada struct       423         433         +2.4%
citm dom            359         361         +0.1%
citm struct         910         886         -2.6%
twitter dom         302         326         +7.9%
twitter struct      611         754         +23%
```

I'd say this is a good result overall, but the citm pessimization bothers me a bit. It seems like the new implementation is slower when most of the strings are very short (less than 16 characters, perhaps?). In practice, this happens when the data contains lots of objects with many keys but small values, e.g. `{"areaId": 205706007, "blockIds": []}` x 10k in citm.

So the questions are:

- Is increased the mental complexity of the core of serde-json fine?
- Is significantly optimizing long strings but pessimizing short strings fine?