@purplesyringa 2024-07-29
Also, just to make sure we're on the same page: as [simd-json](https://docs.rs/simd-json/latest/simd_json/) seems to be serde-compatible, what is the main focus of this crate? Is it supposed to be a simple canonical implementation or are more complicated performance optimizations welcome here?