@purplesyringa 2024-08-11
For completeness, here's the comparison for the benchmark in https://github.com/serde-rs/json/pull/1161#pullrequestreview-2217909786 (5 runs performed):
```
naive   918.20  919.05  921.64  917.83  929.76
#1167   979.75  973.36  983.06  976.38  975.24
old PR  983.56  992.25  975.70  987.34  986.94
new     1342.39 1337.61 1334.17 1333.77 1334.33
```