@purplesyringa 2024-08-11
I think this PR is ready for merging (bar squashing, I'll be happy to squash it if you want).

Unrolling the loop manually twice (this is notably distinct from `u128`) provides very slightly better performance (around +2% compared to +50% of this PR) on the synthetic benchmark from your comment and <1% speed-up on json-benchmark with some regressions. I don't think it's a viable avenue to pursue.

We'd probably have more luck with [wide](https://docs.rs/wide/) for further improvements. I can work on that if you'd like me to, but I think it's better to land this PR in the meantime anyway.

---

> Before merging, I would want to try benchmarking another file that disproportionately contains small string to try to get an upper bound on how much such data would be pessimized, but I expect it won't be much.

Data: `{"ABC": "", "ABC": "", "ABC": "", ...}`.
```
                5 runs (MB/s)
Naive           619.33  626.66  629.83  626.22  626.77
Current PR      504.49  503.71  505.75  505.99  498.94
```
This is a 20% regression.

Data: `["0123456789", "0123456789", "0123456789", ...]`.
```
                5 runs (MB/s)
Naive           980.35  996.23  996.59  980.63  995.74
Current PR      1083.53 1079.53 1066.18 1083.57 1078.90
```
This is a 10% improvement.

Data (strings `"0123"` and `"012"`, chosen randomly): `["0123", "0123", "0123", "012", "012", "0123", "0123", "012", "012", "0123", "012", "012", "012", ...]`.
```
                5 runs (MB/s)
Naive           460.70  460.32  460.03  462.20  461.94
Current PR      647.87  641.29  649.03  647.17  649.25
```

Overall, this PR regresses strings of perfectly predictable very short lengths (the breakeven point seems to be around 6 characters). So the main troublemakers are objects with: a) few, b) short, c) fixed keys and short values.

I'm not sure if this is practical. I cherry-picked some datasets that I expected would be regressed, but none did:
- https://blockchain.info/unconfirmed-transactions?format=json: 25% improvement
- https://data.police.uk/api/leicestershire/neighbourhoods: 12% improvement
- https://raw.githubusercontent.com/Biuni/PokemonGO-Pokedex/master/pokedex.json: 12% improvement

I'm not happy about this, but if this regresses no realistic datasets, we might as well merge it and see who hollers.