@purplesyringa 2024-08-11
I just found a regression. We shouldn't really be looking for short strings as counterexamples, we should be looking for short escape-less intervals, which is really common in Unicode text. For JSON-encoded *War and Peace* in Russian, I get
```
                5 runs (MB/s)
Naive           651.89  648.86  649.74  649.98  651.92
Current PR      536.03  533.37  538.01  541.41  536.62
```